+++ Sources/VSMap/Controllers/MarkerController.swift
@@ -270,7 +270,7 @@ class MarkerController: IMarkerController {
         refreshMarkers()
     }
     
-    func updateLocation(newLocation: CGPoint, precision: Float) {
+    func updateLocation(newLocation: CGPoint, precision: Double) {
 //        let coordinate = newLocation.convertFromMeterToLatLng(converter: converter)
 //        markers.forEach { (key, value) in
 //            if self.setTransparentMarkers(coordinate: coordinate, marker: value) {
